Wire up table deletion/download actions
This commit specializes ItemPanel to TablePanel. This is a bit messy but
it's the best current option. Slot content (such as the table deletion
dialog) can't be given a ref, and that's what is necessary to trigger
table deletion from multiple icon targets.

The abstracted ItemPanel was getting to be too unwieldy in any case, so
this respecializes a TablePanel; the next commit will do the same for
NetworkPanel, and if a mixin or other refactoring mechanism is
warranted, it can be done from that state in a future PR.