Solidify behavior of new panels
This commit adds a lot of logic for handling what happens when the
dialogs associated with the panels are closed in various ways. For
example, if an individual item's delete icon is clicked, then when that
dialog eventually closes, we need to leave any checked items checkboxes
alone afterwards. This is handled by having each dialog's `closed` event
optionally transmit which items were deleted.