Disable nuisance rules
`max-len` has its heart in the right place, but is a dumb rule. The line
length is arbitrary, and there will be times when it hurts readability
to break a line to satisfy it. Disabling this rule moves line length to
the realm of other non-lintable qualities, such as "elegance", choice of
algorithm, etc. If we find that disabling this rule somehow hurts the
codebase, we will re-enable it.

`prefer-default-export` is simply an absurd rule. Just because a module
only exports one name does not, for instance, rule out adding other
exports to that module later. It also hurts the ability to actually
write code: the programmer must now remember whether a module exports
one or more than one name just to write the import statement. A better
approach is to, in general, prefer non-default exports for all modules
unless there is a specific reason not to (e.g., Vue components should
remain default exporters because the whole point of an SFC file is to
export one component).