Use <v-img> in place of <img>
<img> tags are not allowed to self-close, but vim's parser for vue files
doesn't seem to know this, which means the parsing gets all messed up.
Self-closing it fixes the parser, but breaks linting: the linter either
reports it as an error, or it fixes it automatically on commit. Classic
immovable object meets unstoppable force territory.

The solution is to abandon using <img> in place of <v-img> which seems
to be a recommended thing to do anyway: we're using Vuetify after all,
and this same component makes use of <v-img> everywhere else to deal
with images.