Disable linting errors for necessary require() statements
I don't fully understand why these need to be `require()` statements,
but they do. It has to do with Webpack not knowing where to find these
assets at compile time (since they are really runtime assets). I am not
sure if there's a better way to handle this but for now this is the
easiest way forward.