Avoid mutating code in precommit hook
This is a dangerous practice that can result in lost work. For instance,
if you have unsaved changes in some files by accident, and then the
precommit hook runs in fix mode, it may overwrite your unsaved work.
Beyond that, this also just gives away too much control over the
codebase. Perhaps we meant to commit something with `--no-verify`, etc.