Reformat <a-ext> tags to eliminate semantic whitespace
The linter wants to break lines around most elements, including our
homebrew <a-ext>. However, this linebreak introduces a semantic space
which manifests as an ugly space between the end of the slot content and
the beginning of the following content.

The relevant rule has a config option to whitelist an array of tags from
triggering this error; however, that option comes with a *default* array
of elements that there is no easy way to extend rather than replace.
Importing the JSON file from the plugin's own package materials, then
calling `concat()` on it with `['a-ext']` does work, but triggers its
own linting errors within the .eslintrc file. It gets a bit absurd at
that point so the easier thing to do is sprinkle in this one line
disable directives.