Remove the intro popover
This no longer has the right advice--the general user coming to this app
can no longer start out by clicking the "new workspace" button. Guidance
to log in would be a good replacement, but I would prefer to make the UI
itself direct the user to create an account, etc.